Refactor the socketListeners service into a folder structure
