+++ srcs/frontend/static/js/services/index.js
@@ -4,6 +4,6 @@ import i18nextInit from "./i18next.js";
 import navigateTo from "./navigation.js";
 import sendNotification from "./notifications.js";
 import { renderSidebar, renderPage } from "./render.js";
-import socketListeners from "./socket-listeners.js";
+import socketListeners from "./socket-listeners/index.js";
 
 export { doLogout, getCSRFToken, clearTokens, refreshUserID, fetchWithToken, generateSocket, i18nextInit, navigateTo, sendNotification, renderSidebar, renderPage, socketListeners };
