+++ srcs/frontend/static/js/services/socket-listeners/friend-requests.js
@@ -1,34 +1,4 @@
-import { ChatBox } from "/static/js/components/index.js";
 import FriendsSection from "/static/js/components/Sidebar/FriendsSection.js";
-import { sendNotification } from "./index.js";
-
-const onlineUsersSocketListener = () => {
-  // Remove 'online_users' listener
-  SOCKET.off('online_users');
-
-  // Listen to the 'online_users' event
-  SOCKET.on('online_users', (data) => {
-    ONLINE_USERS = data;
-    FriendsSection.updateOnlineStatus();
-  });
-}
-
-const chatMessageSocketListener = () => {
-  // Remove 'chat_message_id' listener
-  SOCKET.off(`chat_message_${USER_ID}`);
-
-  // Listen to the 'chat_message_id' event
-  SOCKET.on(`chat_message_${USER_ID}`, (data) => {
-    // Try to append a message to the ChatBox
-    // but if the ChatBox is not open then send a notification
-    if (!ChatBox.appendMessage(data)) {
-      sendNotification({
-        user: data.sender,
-        body: data.content
-      });
-    }
-  });
-}
 
 const friendAddSocketListener = () => {
   // Remove 'friend_add_id' listener
@@ -70,13 +40,11 @@ const friendCancelSocketListener = () => {
   });
 }
 
-const socketListeners = () => {
-  onlineUsersSocketListener();
-  chatMessageSocketListener();
+const FriendRequests = () => {
   friendAddSocketListener();
   friendRefuseSocketListener();
   friendAcceptSocketListener();
   friendCancelSocketListener();
 }
 
-export default socketListeners;
+export default FriendRequests;
