+++ srcs/frontend/static/js/services/socket-listeners/index.js
@@ -1,0 +1,11 @@
+import chatMessageSocketListener from "./chat-message.js";
+import FriendRequests from "./friend-requests.js";
+import onlineUsersSocketListener from "./online-users.js";
+
+const socketListeners = () => {
+  chatMessageSocketListener();
+  FriendRequests();
+  onlineUsersSocketListener();
+}
+
+export default socketListeners;
